move openai realtime client to seperate file
